feat: extract utils to sub directories, ut in next commit
